core: read the inputstream before the synchronized block
Creating the event reader seems to try to read some bytes from the passed input stream by the looks of it, in order to determine the encoding.
A read on the http input stream causes network access and suddenly all threads are blocked on that first synchronized block.
The hack here is reading the whole input stream before the synchronized block.